requirements.txt
opencv-python 
arrow
rivescript
thefuzz
fuzzywuzzy
thefuzz
humanfriendly
pokebase
pygments
requests
urllib3
qrcode
pyzbar
pillow
pypresense
aiohttp
enolib
enotype